Move the debugging to the test project
