+++ packages/testing-project/favicon.ico
