+++ packages/testing-project/jest.config.cjs
