+++ packages/testing-project/package.json
@@ -15,14 +15,14 @@
     "lint:format": "prettier --check --ignore-path .gitignore .",
     "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
     "start": "lite-server",
-    "test:mutation": "stryker run"
+    "test:mutation": "node ../core/bin/stryker.js run"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@stryker-mutator/core": "^6.4.2",
-    "@stryker-mutator/jest-runner": "^6.4.2",
+    "@stryker-mutator/core": "7.3.0",
+    "@stryker-mutator/jest-runner": "7.3.0",
     "@types/jest": "^29.5.0",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
