+++ packages/testing-project/src/components/place-order.template.js
