+++ packages/testing-project/src/components/robo.component.js
