+++ packages/testing-project/src/pipes/currency.pipe.js
