+++ packages/testing-project/src/pipes/drinks-summary.pipe.js
