+++ packages/testing-project/stryker.conf.json
@@ -1,0 +1,25 @@
+{
+  "$schema": "../core/schema/stryker-schema.json",
+  "packageManager": "npm",
+  "reporters": [
+    "html",
+    "clear-text",
+    "progress"
+  ],
+  "testRunner": "jest",
+  "testRunnerNodeArgs": ["--experimental-vm-modules"],
+  "plugins": [
+    "../jest-runner/dist/src/index.js"
+  ],
+  "jest": {
+    "configFile": "jest.config.cjs"
+  },
+  "coverageAnalysis": "perTest",
+  "mutationLevels": [
+    { 
+      "name": "default",
+      "type": "block", 
+      "mutators": ["ArrayDeclaration"] 
+    }
+  ]
+}
