+++ packages/testing-project/test/pipes/currency.pipe.spec.js
