--- testing-project/stryker.conf.json
@@ -1,11 +1,0 @@
-{
-  "$schema": "./node_modules/@stryker-mutator/core/schema/stryker-schema.json",
-  "_comment": "This config was generated using 'stryker init'. Please take a look at: https://stryker-mutator.io/docs/stryker-js/configuration/ for more information",
-  "packageManager": "npm",
-  "reporters": ["html", "clear-text", "progress"],
-  "testRunner": "jest",
-  "testRunnerNodeArgs": ["--experimental-vm-modules"],
-  "jest": {
-    "configFile": "jest.config.cjs"
-  }
-}
